data_steward/requirements.txt
aniso8601==1.3.0
cachetools==2.0.1
certifi==2017.11.5
chardet==3.0.4
dpath==1.4.2
flask==0.10
Flask-FlatPages==0.6
flask-restful==0.3.5
funcsigs==1.0.2
google_api_python_client==1.6.4
GoogleAppEngineCloudStorageClient==1.9.22.1
httplib2==0.10.3
idna==2.6
itsdangerous==0.24
Jinja2==2.10
Markdown==2.6.9
MarkupSafe==1.0
mock==2.0.0
moz-sql-parser==2.42.19034
oauth2client==4.1.2
pbr==3.1.1
pyasn1==0.4.7
pyasn1_modules==0.2.6
python_dateutil==2.6.1
pytz==2017.3
requests==2.20.0
rsa==3.4.2
six==1.11.0
uritemplate==3.0.0
urllib3==1.22
Werkzeug==0.12.2
google-cloud-logging==1.14.0
